refactor(icoglyphsDB): rename icoGlyphsLibrary.json to icoglyphsDB.json, clean and reorganize /lib/library/external
