+++ src/lib/library/external/getSvgAttributes.js
@@ -1,4 +1,4 @@
-import searchIcoGlyph from '../searchIcoGlyph';
+import searchIcoGlyph from './searchIcoGlyph';
 
 /**
  * @dev In some cases, it is necessary to convert the object to a string.
