+++ src/lib/library/external/searchIcoGlyph.js
@@ -1,11 +1,11 @@
-import icoGlyphsLibrary from '../icoGlyphsLibrary.json';
+import icoglyphsDB from '../icoglyphsDB.json';
 
 /**
  * @param {string} icoGlyphName - The name or alias of the icoGlyph.
  * @returns {object|null} The icoGlyph object, or null if not found.
  */
 function searchIcoGlyph(icoGlyphName) {
-	const ig = icoGlyphsLibrary.svgData.find((entry) => entry.aliases?.includes(icoGlyphName));
+	const ig = icoglyphsDB.svgData.find((entry) => entry.aliases?.includes(icoGlyphName));
 
 	if (!ig) {
 		console.error(`icoGlyph not found: ${icoGlyphName}`);
