+++ src/lib/library/icoglyphsDB.json
