+++ src/lib/library/internal/utilsModJson/utils/fileUtils/pathOfIcoGlyphsLibrary.js
@@ -3,7 +3,7 @@ import { fileURLToPath } from 'url';
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
-// Définir le chemin vers icoGlyphsLibrary.json
-const pathOfIcoGlyphsLibrary = path.resolve(__dirname, '../../../../icoGlyphsLibrary.json');
+// Définir le chemin vers icoglyphsDB.json
+const pathOficoglyphsDB = path.resolve(__dirname, '../../../../icoglyphsDB.json');
 
-export default pathOfIcoGlyphsLibrary;
+export default pathOficoglyphsDB;
